Fleshing out Stored, retrieving now working
The Stored type class is now multi-parameter with a functional
dependency so that the correct monad and a type to identify resources
are both available in the methods.  At the moment this identifying
property is purely pulled from the remaining path of the route, but in
the future could support query strings as well.