Splitting the Stored type class into components
Rather than have all of the methods relating to accessing the resources
from a store be in one type class, each method is now in its own type
class.  This means that certain functionality of a resource can be left
off.  This way, you can create resources which are permanent after their
creation, or adjustable resources that cannot be read.

At this point it's not too clear how to do fully featured options output
from this new interface, as using one of the more specific serving
functions doesn't necessarily address the full feature-set of options
available.