Implementing query search backend
This change replaces FromPath with FromRequest, which allows the
identifying type of a resource to also be parsed from the query string
in the URL.  Implementing these by combining ifTop with serveRoute is
the next step.

The Disabled PutAction has also been removed, as it can be modelled by
specifying an action that has not been implemented, and should not be
able to be turned off if both actions are implemented.